refactor(ui): extract SettingsDialogShell into settings/components/
