+++ packages/ui/src/tabs/settings.tsx
@@ -1,13 +1,12 @@
 /* Settings modal — observer config, sync settings. */
 
 import { render } from "preact";
-import { useCallback, useEffect, useState } from "preact/hooks";
-import { RadixDialog } from "../components/primitives/radix-dialog";
 import * as api from "../lib/api";
 import { $, $button } from "../lib/dom";
 import { showGlobalNotice } from "../lib/notice";
 import { state } from "../lib/state";
 
+import { SettingsDialogShell } from "./settings/components/SettingsDialogShell";
 import { SettingsModalContent } from "./settings/components/SettingsModalContent";
 import { PROTECTED_VIEWER_CONFIG_KEYS } from "./settings/data/constants";
 import { createSettingsEventHandlers } from "./settings/data/event-handlers";
@@ -28,11 +27,7 @@ import type {
 	SettingsRenderState,
 	SettingsTabId,
 } from "./settings/data/types";
-import {
-	mergeOverrideBaseline,
-	persistAdvancedPreference,
-	toProviderList,
-} from "./settings/data/value-helpers";
+import { mergeOverrideBaseline, toProviderList } from "./settings/data/value-helpers";
 
 const getObserverModelHint = (): string =>
 	getObserverModelHintRaw(settingsState.renderState.values, settingsState.envOverrides);
@@ -44,9 +39,6 @@ const getObserverModelTooltip = (): string =>
 	getObserverModelTooltipRaw(settingsState.renderState.values);
 const getObserverModelDescription = (): string =>
 	getObserverModelDescriptionRaw(settingsState.renderState.values);
-
-import { focusSettingsDialog } from "./settings/data/dom";
-import { useHelpTooltip } from "./settings/hooks/use-help-tooltip";
 
 function hideHelpTooltip() {
 	settingsState.controller?.hideTooltip();
@@ -75,7 +67,7 @@ function updateFormState(patch: Partial<SettingsFormState>) {
 function renderSettingsShell() {
 	const mount = $("settingsDialogMount");
 	if (!mount) return;
-	render(<SettingsDialogShell />, mount);
+	render(<SettingsDialogShellBound />, mount);
 	// Lucide icon replacement happens in the shell's open-state effect — the
 	// Dialog renders children only while open, so createIcons() here would
 	// no-op against the unmounted tree.
@@ -89,110 +81,8 @@ function ensureSettingsShell() {
 	settingsState.shellMounted = true;
 }
 
-function SettingsDialogShell() {
-	const [open, setOpen] = useState(settingsState.open);
-	const [activeTab, setActiveTabState] = useState<SettingsTabId>(
-		["observer", "queue", "sync"].includes(settingsState.activeTab)
-			? (settingsState.activeTab as SettingsTabId)
-			: "observer",
-	);
-	const [dirty, setDirtyState] = useState(state.settingsDirty);
-	const [renderState, setRenderStateState] = useState(settingsState.renderState);
-	const [showAdvanced, setShowAdvancedState] = useState(settingsState.showAdvanced);
-	const { tooltipPortal, setTooltip } = useHelpTooltip();
-
-	settingsState.open = open;
-	settingsState.activeTab = activeTab;
-	state.settingsDirty = dirty;
-	settingsState.renderState = renderState;
-	settingsState.showAdvanced = showAdvanced;
-
-	useEffect(() => {
-		settingsState.controller = {
-			hideTooltip: () => {
-				setTooltip({ anchor: null, content: "", visible: false });
-			},
-			setActiveTab: (tab) => {
-				const nextTab = ["observer", "queue", "sync"].includes(tab) ? tab : "observer";
-				settingsState.activeTab = nextTab;
-				setActiveTabState(nextTab);
-			},
-			setDirty: (nextDirty) => {
-				state.settingsDirty = nextDirty;
-				setDirtyState(nextDirty);
-			},
-			setOpen: (nextOpen) => {
-				settingsState.open = nextOpen;
-				setOpen(nextOpen);
-			},
-			setRenderState: (patch) => {
-				const nextState = {
-					...settingsState.renderState,
-					...patch,
-				};
-				settingsState.renderState = nextState;
-				setRenderStateState(nextState);
-			},
-			setShowAdvanced: (nextShowAdvanced) => {
-				settingsState.showAdvanced = nextShowAdvanced;
-				persistAdvancedPreference(nextShowAdvanced);
-				setShowAdvancedState(nextShowAdvanced);
-			},
-		};
-
-		return () => {
-			if (settingsState.controller) {
-				settingsState.controller = null;
-			}
-		};
-	}, []);
-
-	// Radix Dialog mounts its children only while `open` is true, so any
-	// <i data-lucide="..."> stubs inside the dialog need a createIcons pass
-	// every time the modal opens. Running it on the shell mount (before the
-	// children exist) is a no-op for those nodes.
-	useEffect(() => {
-		if (!open) return;
-		const lucide = (globalThis as { lucide?: { createIcons?: () => void } }).lucide;
-		lucide?.createIcons?.();
-	}, [open]);
-
-	const close = useCallback(() => {
-		if (settingsState.startPolling && settingsState.refresh) {
-			closeSettings(settingsState.startPolling, settingsState.refresh);
-		}
-	}, []);
-
-	return (
-		<>
-			<RadixDialog
-				ariaDescribedby="settingsDescription"
-				ariaLabelledby="settingsTitle"
-				contentClassName="modal"
-				contentId="settingsModal"
-				onCloseAutoFocus={(event) => {
-					event.preventDefault();
-				}}
-				onOpenAutoFocus={(event) => {
-					event.preventDefault();
-					focusSettingsDialog();
-				}}
-				onOpenChange={(nextOpen) => {
-					if (nextOpen) {
-						setOpen(true);
-						return;
-					}
-					close();
-				}}
-				open={open}
-				overlayClassName="modal-backdrop"
-				overlayId="settingsBackdrop"
-			>
-				<SettingsDialogContent />
-			</RadixDialog>
-			{tooltipPortal}
-		</>
-	);
+function SettingsDialogShellBound() {
+	return <SettingsDialogShell DialogContent={SettingsDialogContent} onClose={closeSettings} />;
 }
 
 export function isSettingsOpen(): boolean {
